refactor(wl): move filers changes to watchlist_step
