--- watchlist_step/requirements.txt
@@ -1,3 +1,0 @@
-apf_base==1.0.9
-psycopg2
-git+https://github.com/alercebroker/db-plugins.git@2.0.2
