+++ watchlist_step/settings.py
@@ -34,6 +34,8 @@ elif os.getenv("CONSUMER_TOPICS"):
 else:
     raise Exception("Add TOPIC_STRATEGY or CONSUMER_TOPICS")
 
+UPDATE_STRATEGY = os.environ["UPDATE_STRATEGY"]
+
 METRICS_CONFIG = {
     "CLASS": "apf.metrics.KafkaMetricsProducer",
     "EXTRA_METRICS": [
@@ -96,25 +98,13 @@ if os.getenv("METRICS_KAFKA_USERNAME") and os.getenv("METRICS_KAFKA_PASSWORD"):
 
 ## Step Configuratiom
 STEP_CONFIG = {
-    "alert_db_config": {
-        "SQL": {
-            "ENGINE": "postgresql",
-            "HOST": os.environ["ALERTS_DB_HOST"],
-            "USER": os.environ["ALERTS_DB_USER"],
-            "PASSWORD": os.environ["ALERTS_DB_PASSWORD"],
-            "PORT": 5432,  # postgresql tipically runs on port 5432. Notice that we use an int here.
-            "DB_NAME": os.environ["ALERTS_DB_NAME"],
-        },
-    },
-    "users_db_config": {
-        "SQL": {
-            "ENGINE": "postgresql",
-            "HOST": os.environ["USERS_DB_HOST"],
-            "USER": os.environ["USERS_DB_USER"],
-            "PASSWORD": os.environ["USERS_DB_PASSWORD"],
-            "PORT": 5432,  # postgresql tipically runs on port 5432. Notice that we use an int here.
-            "DB_NAME": os.environ["USERS_DB_NAME"],
-        }
+    "PSQL_CONFIG": {
+        "ENGINE": "postgresql",
+        "HOST": os.environ["USERS_DB_HOST"],
+        "USER": os.environ["USERS_DB_USER"],
+        "PASSWORD": os.environ["USERS_DB_PASSWORD"],
+        "PORT": 5432,  # postgresql tipically runs on port 5432. Notice that we use an int here.
+        "DB_NAME": os.environ["USERS_DB_NAME"],
     },
     "METRICS_CONFIG": METRICS_CONFIG,
 }
