+++ watchlist_step/tests/integration/docker-compose.yml
@@ -24,16 +24,7 @@ services:
       - KAFKA_JMX_PORT=9101
     ports:
       - "9094:9094"
-
-  alerts_db:
-    image: postgres
-    environment:
-      - POSTGRES_USER=postgres
-      - POSTGRES_DB=postgres
-      - POSTGRES_PASSWORD=postgres
-    ports:
-      - "5432:5432"
-
+  
   users_db:
     build: docker/postgres
     environment:
