+++ watchlist_step/watchlist_step/strategies/sorting_hat.py
@@ -1,6 +1,6 @@
 from typing import List, Tuple
 
-from wl_match_step.strategies.base import BaseStrategy
+from watchlist_step.strategies.base import BaseStrategy
 
 
 class SortingHatStrategy(BaseStrategy):
