--- wl_match_step/.dockerignore
@@ -1,23 +1,0 @@
-.git
-.github
-.dockerignore
-.gitignore
-
-.idea
-.vscode
-
-__pycache__/
-*.py[cod]
-*$py.class
-*.so
-htmlcov/
-.coverage
-.coverage.*
-.pytest_cache/
-.venv
-venv
-
-.DS_Store
-.AppleDouble
-.LSOverride
-._*
