--- wl_match_step/README.md
@@ -1,1 +1,0 @@
-# Watchlist Step
