--- wl_match_step/pyproject.toml
@@ -1,31 +1,0 @@
-[tool.poetry]
-name = "wl-match-step"
-version = "23.12.26a51"
-description = "Watchlist Step"
-authors = []
-readme = "README.md"
-packages = [{ include = "wl_match_step" }]
-
-[tool.poetry.scripts]
-step = { callable = "scripts.run_step:step" }
-
-[tool.poetry.dependencies]
-python = "~3.11"
-apf-base = "^1.0.9"
-pydantic = "^2.6.4"
-psycopg = "^3.1.18"
-
-[tool.poetry.group.test.dependencies]
-pytest = "^7.4.0"
-pytest-docker = "^1.0.1"
-pytest-cov = "^4.1.0"
-
-[tool.poetry.group.dev.dependencies]
-black = "^23.0.0"
-
-[tool.black]
-line-length = 88
-
-[build-system]
-requires = ["poetry-core"]
-build-backend = "poetry.core.masonry.api"
