--- wl_match_step/settings.py
@@ -1,110 +1,0 @@
-##################################################
-#       watchlist_step   Settings File
-##################################################
-import os
-
-## Set the global logging level to debug
-# LOGGING_DEBUG = True
-
-## Consumer configuration
-### Each consumer has different parameters and can be found in the documentation
-CONSUMER_CONFIG = {
-    "PARAMS": {
-        "bootstrap.servers": os.environ["CONSUMER_SERVER"],
-        "group.id": os.environ["CONSUMER_GROUP_ID"],
-        "auto.offset.reset": "beginning",
-        "max.poll.interval.ms": 3600000,
-        "enable.partition.eof": os.getenv("ENABLE_PARTITION_EOF", False),
-    },
-    "consume.timeout": int(os.getenv("CONSUME_TIMEOUT", 10)),
-    "consume.messages": int(os.getenv("CONSUME_MESSAGES", 1000)),
-}
-
-if os.getenv("TOPIC_STRATEGY_FORMAT"):
-    CONSUMER_CONFIG["TOPIC_STRATEGY"] = {
-        "CLASS": "apf.core.topic_management.DailyTopicStrategy",
-        "PARAMS": {
-            "topic_format": os.environ["TOPIC_STRATEGY_FORMAT"].strip().split(","),
-            "date_format": "%Y%m%d",
-            "change_hour": 23,
-        },
-    }
-elif os.getenv("CONSUMER_TOPICS"):
-    CONSUMER_CONFIG["TOPICS"] = os.environ["CONSUMER_TOPICS"].strip().split(",")
-else:
-    raise Exception("Add TOPIC_STRATEGY or CONSUMER_TOPICS")
-
-UPDATE_STRATEGY = os.environ["UPDATE_STRATEGY"]
-
-METRICS_CONFIG = {
-    "CLASS": "apf.metrics.KafkaMetricsProducer",
-    "EXTRA_METRICS": [
-        {"key": "candid", "format": lambda x: str(x)},
-        {"key": "objectId", "alias": "oid"},
-    ],
-    "PARAMS": {
-        "PARAMS": {
-            "bootstrap.servers": os.environ["METRICS_HOST"],
-        },
-        "TOPIC": os.environ["METRICS_TOPIC"],
-        "SCHEMA": {
-            "$schema": "http://json-schema.org/draft-07/schema",
-            "$id": "http://example.com/example.json",
-            "type": "object",
-            "title": "The root schema",
-            "description": "The root schema comprises the entire JSON document.",
-            "default": {},
-            "examples": [
-                {"timestamp_sent": "2020-09-01", "timestamp_received": "2020-09-01"}
-            ],
-            "required": ["timestamp_sent", "timestamp_received"],
-            "properties": {
-                "timestamp_sent": {
-                    "$id": "#/properties/timestamp_sent",
-                    "type": "string",
-                    "title": "The timestamp_sent schema",
-                    "description": "Timestamp sent refers to the time at which a message is sent.",
-                    "default": "",
-                    "examples": ["2020-09-01"],
-                },
-                "timestamp_received": {
-                    "$id": "#/properties/timestamp_received",
-                    "type": "string",
-                    "title": "The timestamp_received schema",
-                    "description": "Timestamp received refers to the time at which a message is received.",
-                    "default": "",
-                    "examples": ["2020-09-01"],
-                },
-            },
-            "additionalProperties": True,
-        },
-    },
-}
-
-if os.getenv("CONSUMER_KAFKA_USERNAME") and os.getenv("CONSUMER_KAFKA_PASSWORD"):
-    CONSUMER_CONFIG["PARAMS"]["security.protocol"] = "SASL_SSL"
-    CONSUMER_CONFIG["PARAMS"]["sasl.mechanism"] = "SCRAM-SHA-512"
-    CONSUMER_CONFIG["PARAMS"]["sasl.username"] = os.getenv("CONSUMER_KAFKA_USERNAME")
-    CONSUMER_CONFIG["PARAMS"]["sasl.password"] = os.getenv("CONSUMER_KAFKA_PASSWORD")
-if os.getenv("METRICS_KAFKA_USERNAME") and os.getenv("METRICS_KAFKA_PASSWORD"):
-    METRICS_CONFIG["PARAMS"]["PARAMS"]["security.protocol"] = "SASL_SSL"
-    METRICS_CONFIG["PARAMS"]["PARAMS"]["sasl.mechanism"] = "SCRAM-SHA-512"
-    METRICS_CONFIG["PARAMS"]["PARAMS"]["sasl.username"] = os.getenv(
-        "METRICS_KAFKA_USERNAME"
-    )
-    METRICS_CONFIG["PARAMS"]["PARAMS"]["sasl.password"] = os.getenv(
-        "METRICS_KAFKA_PASSWORD"
-    )
-
-## Step Configuratiom
-STEP_CONFIG = {
-    "PSQL_CONFIG": {
-        "ENGINE": "postgresql",
-        "HOST": os.environ["USERS_DB_HOST"],
-        "USER": os.environ["USERS_DB_USER"],
-        "PASSWORD": os.environ["USERS_DB_PASSWORD"],
-        "PORT": 5432,  # postgresql tipically runs on port 5432. Notice that we use an int here.
-        "DB_NAME": os.environ["USERS_DB_NAME"],
-    },
-    "METRICS_CONFIG": METRICS_CONFIG,
-}
