--- wl_match_step/tests/integration/samples/targets/data/targets.csv
@@ -1,6 +1,0 @@
-1;252.678866288639;53.3452115857332;5;{"""fields""":{"""sorting_hat""":["""mag""", """mjd"""]},"""query""":"""TEST""","""filters""":[{"""type""":"""constant""","""params""":{"""field""":"""mag""","""constant""":10,"""op""": """less"""}}]}
-2;316.427789720271;-57.7893762583343;0.000568402267547;{"""fields""":{"""sorting_hat""":["""mag"""]},"""query""":"""TEST""","""filters""":[{"""type""":"""constant""","""params""":{"""field""":"""mag""","""constant""":10,"""op""": """less"""}}]}
-3;237.397938935363;85.4059039252133;0.001111297038408;{"""fields""":{"""sorting_hat""":["""mag"""]},"""query""":"""TEST""","""filters""":[{"""type""":"""constant""","""params""":{"""field""":"""mag""","""constant""":10,"""op""": """less"""}}]}
-4;52.2711053179025;-40.4163467436118;0.0006527289517;{"""fields""":{"""sorting_hat""":["""mag"""]},"""query""":"""TEST""","""filters""":[{"""type""":"""constant""","""params""":{"""field""":"""mag""","""constant""":10,"""op""": """less"""}}]}
-5;162.408378852195;-28.3087014352101;0.005093634566095;{"""fields""":{"""sorting_hat""":["""mag"""]},"""query""":"""TEST""","""filters""":[{"""type""":"""constant""","""params""":{"""field""":"""mag""","""constant""":10,"""op""": """less"""}}]}
-6;258.820792113938;-53.0924048292811;0.000540467126969;{"""fields""":{"""sorting_hat""":["""mag"""]},"""query""":"""TEST""","""filters""":[{"""type""":"""constant""","""params""":{"""field""":"""mag""","""constant""":10,"""op""": """less"""}}]}
