--- wl_match_step/tests/unittest/test_step.py
@@ -1,75 +1,0 @@
-import datetime
-from unittest import mock
-
-import psycopg
-import pytest
-
-from wl_match_step.step import WatchlistMatchStep
-
-
-@pytest.fixture
-def wl_step():
-    consumer = mock.MagicMock()
-    strategy_name = "SortingHat"
-
-    config = {
-        "PSQL_CONFIG": {
-            "ENGINE": "postgresql",
-            "HOST": "localhost",
-            "USER": "postgres",
-            "PASSWORD": "password",
-            "PORT": 5433,
-            "DB_NAME": "postgres",
-        }
-    }
-    return WatchlistMatchStep(
-        consumer=consumer,
-        strategy_name=strategy_name,
-        config=config,
-    )
-
-
-@mock.patch("wl_match_step.step.datetime.datetime")
-@mock.patch("psycopg.connect")
-def test_should_insert_matches(
-    connect_mock: mock.MagicMock,
-    datetime_mock: mock.MagicMock,
-    wl_step: WatchlistMatchStep,
-):
-    datetime_mock.now.return_value = "date"
-
-    cursor_mock: mock.MagicMock = (
-        connect_mock.return_value.__enter__.return_value.cursor.return_value.__enter__.return_value
-    )
-    executemany_mock: mock.MagicMock = cursor_mock.executemany
-
-    wl_step.insert_matches([(0, 1, 2), (3, 4, 5)])
-
-    executemany_mock.assert_called_once()
-
-    assert (
-        len(executemany_mock.call_args.args[1]) == 2
-    ), "executemany should be called with two matches"
-
-    assert (
-        len(executemany_mock.call_args.args[1][0]) == 5
-    ), "executemany should be called with tuples of length 5"
-
-
-@mock.patch("psycopg.connect")
-def test_should_create_matches(
-    connect_mock: mock.MagicMock, wl_step: WatchlistMatchStep
-):
-    cursor_mock: mock.MagicMock = (
-        connect_mock.return_value.__enter__.return_value.cursor.return_value.__enter__.return_value
-    )
-    execute_mock: mock.MagicMock = cursor_mock.execute
-
-    wl_step.match_user_targets(
-        [("ra1", "dec1", "oid1", "candid1"), ("ra2", "dec2", "oid2", "candid2")]
-    )
-
-    execute_mock.assert_called_once()
-    assert (
-        len(execute_mock.call_args.args[1]) == 8
-    ), "execute should be called with a flattened list"
