--- wl_match_step/wl_match_step/__init__.py
@@ -1,2 +1,0 @@
-from . import step
-from .step import *
